Added ASM Source
How to make:
Simply run nasm -f bin os.asm -o os.img
Designed to be run raw. Should work on Hard Disk or floppy.
Note that some bios vendors do trickery, common in UEFI vendors, which may cause CobaltOS to either not be recognized as an OS, or be frozen in a black screen. This was tested in QEMU, so I suggest you run it there.